overview

+++ src/main/scala/io/turntabl/grades/Grade.scala
@@ -1,3 +1,3 @@
 package io.turntabl.grades
 
-case class Grade(subject: String, grad: Int)
+case class Grade(subject: String, grad: Double)
